+++ mmq.c
@@ -12,12 +12,12 @@
 
 void print_residue(points_t *points, Vector *b)
 {
-    int num_points = points->num_of_points;
+    long long int num_points = points->num_of_points;
     
-    for (int i = 0; i < num_points; i++)
+    for (long long int i = 0; i < num_points; i++)
     {
         Interval sum = interval(0.0);
-        for (int j = 0; j < b->size; j++)
+        for (long long int j = 0; j < b->size; j++)
         {
             sum = interval_sum(sum, interval_mul(b->data[j], interval_pow(interval(points->points[i].x), j)));
         }
@@ -27,23 +27,23 @@ void print_residue(points_t *points, Vector *b)
     printf("\n");
 }
 
-void generate_matrix(Matrix *A, Vector *b, points_t *points, unsigned int order)
+void generate_matrix(Matrix *A, Vector *b, points_t *points, unsigned long long int order)
 {
     // Fill the matrix and the vector
-    for (int i = 0; i < A->size; i++)
+    for (long long int i = 0; i < A->size; i++)
     {
-        for (int j = 0; j < A->size; j++)
+        for (long long int j = 0; j < A->size; j++)
         {
             A->data[i][j] = interval(0.0);
-            for (int k = 0; k < points->num_of_points; k++)
+            for (long long int k = 0; k < points->num_of_points; k++)
                 A->data[i][j] = interval_sum(A->data[i][j], interval_pow(interval(points->points[k].x), i + j));
         }
     }
 
-    for (int i = 0; i < b->size; i++)
+    for (long long int i = 0; i < b->size; i++)
     {
         b->data[i] = interval(0.0);
-        for (int k = 0; k < points->num_of_points; k++)
+        for (long long int k = 0; k < points->num_of_points; k++)
             b->data[i] = interval_sum(b->data[i], interval_mul(interval_pow(interval(points->points[k].x), i), interval(points->points[k].y)));
     }
 }
